+++ server/impl/src/test/java/com/walmartlabs/concord/server/TestObjectMapper.java
@@ -1,4 +1,4 @@
-package com.walmartlabs.concord.server.process.queue;
+package com.walmartlabs.concord.server;
 
 /*-
  * *****
@@ -9,9 +9,9 @@ package com.walmartlabs.concord.server.process.queue;
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
- *
+ * 
  *      http://www.apache.org/licenses/LICENSE-2.0
- *
+ * 
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
@@ -20,16 +20,18 @@ package com.walmartlabs.concord.server.process.queue;
  * =====
  */
 
-import com.walmartlabs.concord.server.process.ProcessStatus;
+import com.fasterxml.jackson.databind.ObjectMapper;
 
-import java.util.UUID;
+public final class TestObjectMapper {
 
-public interface ProcessWaitHandler<T extends AbstractWaitCondition> {
+    public static final ObjectMapper INSTANCE = createObjectMapper();
 
-    WaitType getType();
+    private static ObjectMapper createObjectMapper() {
+        ObjectMapper om = new ObjectMapper();
+        new ObjectMapperInitializer(() -> om);
+        return om;
+    }
 
-    /**
-     * @return {@code null} if the specified process doesn't have any wait conditions.
-     */
-    T process(UUID instanceId, ProcessStatus status, T waits);
+    private TestObjectMapper() {
+    }
 }